divisors sum factorize

+++ src/divisors_from_prime_factors.rs
@@ -1,4 +1,4 @@
 // TODO:
-pub fn divisors_from_prime_factors(prime_factors: &[(u64, usize)]) -> Vec<u64> {
+pub fn divisors_from_prime_factors(prime_factors: &[(u64, u8)]) -> Vec<u64> {
     vec![]
 }
